Add ui options for cli and streamlit
