+++ improve_my_cv/cv_improve.py
@@ -1,12 +1,12 @@
 import json
 
 from improve_my_cv.llm_handlers.base_handler import LLMHandler
-from improve_my_cv.llm_handlers.ollama import HandleOllama
 from improve_my_cv.log_config import logger
 from improve_my_cv.prompt_creator import PromptCreator
 
-DEFAULT_MODEL = 'llama3'
-DEFAULT_LLMHANDLER = HandleOllama()
+
+class InvalidModelException(Exception):
+    pass
 
 
 class InvalidResponseException(Exception):
@@ -26,12 +26,20 @@ class ImproveMyCV:
         self.original_resume = original_resume
         self.prompt = PromptCreator(json_resume=json.dumps(self.original_resume), job_description=job_description).create_prompt()
         self.improved_resume: dict = None
+        self.model = None
+        self.llm_handler = None
         logger.debug(f'prompt: {self.prompt}')
 
-    def improve_cv(self, model: str = DEFAULT_MODEL, llm_handler: LLMHandler = DEFAULT_LLMHANDLER) -> dict:
-        logger.info(f'Attempting to improve resume with model {model}')
-        llm_handler.generate(prompt=self.prompt, model=model)
-        model_response = llm_handler.standardize_response()
+    def llm_setup(self, model: str, llm_handler: LLMHandler) -> None:
+        if self.model is None or self.llm_handler is None:
+            raise InvalidModelException(f'You must select a model and llm_handler by running {__name__}')
+        self.model = model
+        self.llm_handler = llm_handler
+
+    def improve_cv(self) -> dict:
+        logger.info(f'Attempting to improve resume with model {self.model}')
+        self.llm_handler.generate(prompt=self.prompt, model=self.model)
+        model_response = self.llm_handler.standardize_response()
 
         try:
             self.improved_resume = json.loads(model_response.text)
